Collapses meeting minutes and board transition
Gets rid of the shallow minutes page in favor for a link dropdown that is
initially hidden until a mouseover. This does introduce the problem that
the borders lose their eveness when the link is expanded. I don't think
this is such an issue as archive will hardly be expanded, so hardly anyone
will notice.

Our site now uses javascript.

Closes #26